# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Baroneza",
    "description": "A release train to automate boring tasks",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Your basic localhost"
    }
  ],
  "components": {},
  "paths": {
    "/startTrain": {
      "post": {
        "description": "This endpoint guesses the next release of the repository, creates a PR, reads the commits, parses the commit messages, tags the tickets and creates a changelog",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "repository": {
                    "type": "string",
                    "example": "baroneza-test"
                  },
                  "baseBranch": {
                    "type": "string",
                    "example": "develop"
                  },
                  "targetBranch": {
                    "type": "string",
                    "example": "master"
                  },
                  "channel": {
                    "type": "string",
                    "example": "automation"
                  },
                  "jiraTagSuffix": {
                    "type": "string",
                    "example": "-backend",
                    "description": "This field will be concatenated as a suffix of the guessed next release"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/tagRelease": {
      "post": {
        "description": "This endpoint reads an open pull request or git's history based on a commit's SHA, parses the commit messages for Jira tickets and then add-version on then.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "reference": {
                    "type": "integer | string",
                    "description": "The number of the Pull Request or the SHA of the commit to read from",
                    "example": 1
                  },
                  "tag": {
                    "type": "string",
                    "description": "The tag on Jira that you want to add to the tickets",
                    "example": "1.0.0"
                  },
                  "projectKeys": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "example": ["ABC"]
                  },
                  "repository": {
                    "type": "string",
                    "description": "The Github repository",
                    "example": "baroneza-test"
                  },
                  "jiraTagSuffix": {
                    "type": "string",
                    "example": "-backend",
                    "description": "This field will be concatenated as a suffix of the guessed next release"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/createRelease": {
      "post": {
        "description": "This endpoint creates a branch, parses commits, creates a changelog",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "branchName": {
                    "type": "string",
                    "example": "rc-1.0.0"
                  },
                  "referenceBranch": {
                    "type": "string",
                    "example": "develop"
                  },
                  "title": {
                    "type": "string",
                    "example": "Awesome Release Candidate"
                  },
                  "targetBranch": {
                    "type": "string",
                    "example": "release-candidate"
                  },
                  "projectTag": {
                    "type": "string",
                    "example": "1.0.0"
                  },
                  "projectKeys": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "example": ["ABC"]
                  },
                  "repository": {
                    "type": "string",
                    "example": "baroneza-test"
                  },
                  "channel": {
                    "type": "string",
                    "example": "automation"
                  },
                  "jiraTagSuffix": {
                    "type": "string",
                    "example": "-backend",
                    "description": "This field will be concatenated as a suffix of the guessed next release"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/updateRelease": {
      "post": {
        "description": "This endpoint parses all commits of a branch, looks for Jira tickets, notify the ones that were not tagged yet, update milestone, Jira Release, PR description and Github release description",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "channel": {
                    "type": "string",
                    "example": "automation"
                  },
                  "fromVersion": {
                    "type": "string",
                    "example": "1.0.0"
                  },
                  "jiraSuffix": {
                    "type": "string",
                    "example": "-backend"
                  },
                  "projectKeys": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "example": ["ABC"]
                  },
                  "pullRequestNumber": {
                    "type": "number",
                    "example": 57
                  },
                  "repository": {
                    "type": "string",
                    "example": "baroneza-test"
                  },
                  "title": {
                    "type": "string",
                    "example": "Baroneza Staging Nightly Build"
                  },
                  "toVersion": {
                    "type": "string",
                    "example": "2.0.0"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/releaseVersion": {
      "post": {
        "description": "This endpoint releases a version in JIRA",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "projectTag": {
                    "type": "string",
                    "example": "1.0.0"
                  },
                  "projectKeys": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "example": ["ABC"]
                  },
                  "jiraTagSuffix": {
                    "type": "string",
                    "example": "-backend",
                    "description": "This field will be concatenated as a suffix of the guessed next release"
                  },
                  "releaseDate": {
                    "type": "string",
                    "example": "(Optional) Date of the release, expressed in ISO 8601 format (yyyy-mm-dd). Defaults to current date."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  }
}